requirements.txt
PyYAML==5.4.1
lxml==4.6.3